Moved creation and setting of SSLContext to ElasticSearchDao
