+++ src/main/java/jenkins/plugins/logstash/persistence/ElasticSearchDao.java
@@ -47,10 +47,22 @@ import java.net.URISyntaxException;
 
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
+import java.security.KeyManagementException;
+import java.security.KeyStore;
+import java.security.KeyStoreException;
+import java.security.NoSuchAlgorithmException;
+import java.security.cert.Certificate;
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
 
 import com.google.common.collect.Range;
 
 import jenkins.plugins.logstash.configuration.ElasticSearch;
+
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.TrustManager;
+import javax.net.ssl.TrustManagerFactory;
+import javax.net.ssl.X509TrustManager;
 
 
 /**
@@ -69,12 +81,15 @@ public class ElasticSearchDao extends AbstractLogstashIndexerDao {
   private String username;
   private String password;
   private String mimeType;
-  
+  private KeyStore customKeyStore;
+
+  //primary constructor used by indexer factory
   public ElasticSearchDao(URI uri, String username, String password) {
     this(null, uri, username, password);
   }
 
-  public ElasticSearchDao(HttpClientBuilder factory, URI uri, String username, String password) {
+  // Factored for unit testing
+  ElasticSearchDao(HttpClientBuilder factory, URI uri, String username, String password) {
 
     if (uri == null)
     {
@@ -146,6 +161,14 @@ public class ElasticSearchDao extends AbstractLogstashIndexerDao {
   public void setMimeType(String mimeType) {
     this.mimeType = mimeType;
   }
+
+  public KeyStore getCustomKeyStore() {
+    return this.customKeyStore;
+  }
+
+  public void setCustomKeyStore(KeyStore customKeyStore) {
+    this.customKeyStore = customKeyStore;
+  }
   
   String getAuth()
   {
@@ -172,7 +195,20 @@ public class ElasticSearchDao extends AbstractLogstashIndexerDao {
     CloseableHttpResponse response = null;
     HttpPost post = getHttpPost(data);
 
+    if (customKeyStore != null) {
+      try {
+        String alias;
+        alias = customKeyStore.aliases().nextElement();
+        X509Certificate certificate = (X509Certificate) customKeyStore.getCertificate(alias);
+        if (certificate != null)
+          clientBuilder.setSslcontext(ElasticSearchDao.createSSLContext(alias, certificate));
+      } catch (KeyStoreException | CertificateException | NoSuchAlgorithmException | KeyManagementException e) {
+        e.printStackTrace();
+      }
+    }
+
     try {
+
       httpClient = clientBuilder.build();
       response = httpClient.execute(post);
 
@@ -217,6 +253,77 @@ public class ElasticSearchDao extends AbstractLogstashIndexerDao {
         stream.close();
       }
     }
+  }
+
+  public static SSLContext createSSLContext(String alias, X509Certificate certificate)
+          throws CertificateException, NoSuchAlgorithmException, IOException, KeyStoreException, KeyManagementException
+  {
+    // Step 1: Get all defaults
+    TrustManagerFactory tmf = TrustManagerFactory
+        .getInstance(TrustManagerFactory.getDefaultAlgorithm());
+    // Using null here initialises the TMF with the default trust store.
+    tmf.init((KeyStore) null);
+
+    // Get hold of the default trust manager
+    X509TrustManager defaultTM = null;
+    for (TrustManager tm : tmf.getTrustManagers()) {
+      if (tm instanceof X509TrustManager) {
+        defaultTM = (X509TrustManager) tm;
+        break;
+      }
+    }
+
+    // Step 2: Add custom cert to keystore
+    KeyStore ks = KeyStore.getInstance(KeyStore.getDefaultType());
+    ks.load(null, null);
+    ks.setEntry(alias, new KeyStore.TrustedCertificateEntry(certificate), null);
+
+    // Create TMF with our custom cert's KS
+    tmf = TrustManagerFactory
+        .getInstance(TrustManagerFactory.getDefaultAlgorithm());
+    tmf.init(ks);
+
+    // Get hold of the custom trust manager
+    X509TrustManager customTM = null;
+    for (TrustManager tm : tmf.getTrustManagers()) {
+      if (tm instanceof X509TrustManager) {
+        customTM = (X509TrustManager) tm;
+        break;
+      }
+    }
+
+    // Step 3: Wrap it in our own class.
+    final X509TrustManager finalDefaultTM = defaultTM;
+    final X509TrustManager finalCustomTM = customTM;
+    X509TrustManager combinedTM = new X509TrustManager() {
+      @Override
+      public X509Certificate[] getAcceptedIssuers() {
+        return finalDefaultTM.getAcceptedIssuers();
+      }
+
+      @Override
+      public void checkServerTrusted(X509Certificate[] chain,
+                       String authType) throws CertificateException {
+        try {
+          finalCustomTM.checkServerTrusted(chain, authType);
+        } catch (CertificateException e) {
+          // This will throw another CertificateException if this fails too.
+          finalDefaultTM.checkServerTrusted(chain, authType);
+        }
+      }
+
+      @Override
+      public void checkClientTrusted(X509Certificate[] chain,
+                       String authType) throws CertificateException {
+        finalDefaultTM.checkClientTrusted(chain, authType);
+      }
+    };
+
+    // Step 4: Finally, create SSLContext based off of this combined TM
+    SSLContext sslContext = SSLContext.getInstance("TLS");
+    sslContext.init(null, new TrustManager[]{combinedTM}, null);
+
+    return sslContext;
   }
 
   @Override
